Aligned the "Minute" members of intraday and historical data points.
They appear to both be used the same way, as a string field that tells
us the minute of the data point in the time zone of the exchange.

I also removed a helper function that attempts to get the time from the
historical data point by combining the date and minute fields. The
problem is that this results in the wrong time because the time objects
are in UTC, so to do this right you'd need to know the time zone which
the "minute" value is in. I think we should leave it up to the consumer
to write their methods to convert times according to their special
circumstances.